refactor(divider): script setup
