+++ src/packages/__VUE/divider/doc.en-US.md
@@ -65,6 +65,17 @@ User can custom divider style with style attribute.
 | --- | --- |
 | default | Default slot, when `direction` = `horizontal` |
 
+### Types
+
+The component exports the following type definitions:
+
+```js
+import type {
+  DividerPosition,
+  DividerDirection
+} from '@nutui/nutui';
+```
+
 ## Theming
 
 ### CSS Variables
