+++ src/packages/__VUE/divider/index.ts
@@ -1,0 +1,11 @@
+import Divider from './divider.vue';
+import type { ComponentPublicInstance } from 'vue';
+import { withInstall } from '@/packages/utils';
+
+withInstall(Divider);
+
+export type { DividerPosition, DividerDirection } from './types';
+
+export type DividerInstance = ComponentPublicInstance & InstanceType<typeof Divider>;
+
+export { Divider, Divider as default };
